수정과 삭제 API 예외처리 완료

+++ library-app/src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -1,16 +1,13 @@
 package com.group.libraryapp.controller.user;
 
-import com.group.libraryapp.domain.user.User;
-import com.group.libraryapp.dto.request.UserUpdateRequest;
+import com.group.libraryapp.dto.user.request.UserUpdateRequest;
 import com.group.libraryapp.dto.user.request.UserCreateRequest;
 import com.group.libraryapp.dto.user.response.UserResponse;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
+
 import org.springframework.web.bind.annotation.*;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
+
 import java.util.List;
 
 @RestController
@@ -42,12 +39,27 @@ public class UserController {
 
     @PutMapping("/user")
     public void updateUser(@RequestBody UserUpdateRequest request) {
+        String readSql = "SELECT * FROM user WHERE id =?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, request.getId()).isEmpty();
+
+        if (isUserNotExist) {
+            throw new IllegalArgumentException("유저가 없다 이친구야...");
+        }
+
         String sql = "UPDATE user SET name =? WHERE id =?";
         jdbcTemplate.update(sql, request.getName(), request.getId());
     }
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name) {
+
+        String readSql = "SELECT * FROM user WHERE name =?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, name).isEmpty();
+
+        if (isUserNotExist) {
+            throw new IllegalArgumentException("유저가 없다 이친구야...");
+        }
+
         String sql = "DELETE FROM user WHERE name =?";
         jdbcTemplate.update(sql, name);
     }
